multi-agent-rag/backend/requirements.txt
flask==3.1.2
sqlalchemy==2.0.43
python-dotenv==1.1.1
sentence-transformers==5.1.1
faiss-cpu==1.12.0
PyMuPDF==1.26.4
requests==2.32.5
arxiv==2.2.0
reportlab==4.0.0


flask 
sqlalchemy 
python-dotenv
sentence-transformers
faiss-cpu
pymupdf  
requests
arxiv
reportlab
celery
redis